requirements.txt
accelerate==1.8.1
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
breadability==0.1.20
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.11
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.5
cycler==0.12.1
dashscope==1.23.8
dataclasses-json==0.6.7
datasets==2.21.0
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
docx2txt==0.8
durationpy==0.10
editdistance==0.8.1
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
faiss-cpu==1.8.0.post1
fastapi==0.112.0
ffmpy==0.6.0
filelock==3.13.1
filetype==1.2.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.6.1
funasr==1.2.6
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
gradio==4.32.0
gradio_client==0.17.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.1
huggingface-hub==0.34.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inquirerpy==0.3.4
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jusText==3.0.2
kaldiio==2.18.1
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.43
langchain-huggingface==0.3.0
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.147
lazy_loader==0.4
librosa==0.11.0
llama-index-core==0.10.67
llama-index-embeddings-dashscope==0.1.4
llama-index-embeddings-huggingface==0.2.3
llama-index-embeddings-langchain==0.1.2
llama-index-postprocessor-dashscope-rerank-custom==0.1.0
llama-index-readers-file==0.1.33
llama-index-vector-stores-faiss==0.1.2
llvmlite==0.44.0
lxml==6.0.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.3
mdurl==0.1.2
mmh3==5.2.0
modelscope==1.18.0
mpmath==1.3.0
msgpack==1.1.1
msvc_runtime==14.42.34433
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
narwhals==1.46.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
noisereduce==3.0.3
numba==0.61.2
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
openai==1.93.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.10.18
oss2==2.19.0
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20250506
pfzy==0.3.4
pi_heif==1.0.0
pikepdf==9.10.2
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocotools==2.0.10
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.2
PyMySQL==1.1.1
pynndescent==0.5.13
pyparsing==3.2.3
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytorch-wpe==0.0.1
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.2
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==2.0.41
starlette==0.37.2
striprtf==0.0.26
sumy==0.11.0
sympy==1.13.3
tenacity==8.5.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.19
tokenizers==0.21.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.7.1+cu118
torch-complex==0.4.4
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.51.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
umap-learn==0.5.9.post2
unstructured==0.18.3
unstructured-client==0.38.1
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
